Add the combo box tree image example
The tree example, which shows category titles within a combo box, makes the titles lighter. A normal method for categorization is to make a title bigger or bolder via standard text hierarchy. Plus it's best not to make text lighter and harder to read, and it can be mistaken for disabled text which could confuse the uer.